Expand path parameter regex
In 0cd4f22ee758500942ef296c274c3872ffc65fe4, a change was made to
validate that path parameters are actually in the path. Unfortunately,
the regex to find the path parameters was too restrictive and only
looked for alphanumeric parameter names. The OpenAPI3 spec is ambiguous
as to what characters are allowed in a path parameter name, as
https://github.com/OAI/OpenAPI-Specification/issues/2119 describes in
detail. Until that question is unanswered, this commit takes a
conservative approach by extending the regex to match the "unreserved"
characters of RFC 3986.