Better errors when a single type is required
In making #45, I noticed the validation error was not helpful, and would
be better if it listed out the required fields in the case that only a
single type is allowed.  The old error is still raised in cases where
there are multiple possibilities for a value, since I can't know which
required fields to display, but in the simple case of one specific type
being needed, we can be more helpful.