bug: Resolve references in components first
Related to https://github.com/linode/linode-cli/pull/423

In the above, it was discovered that a reference in a `requestBody`
schema that references a reference in `components.schema` will see the
`Reference` object once resolution is complete, which is unexpected; it
should see the final `Schema`.

This change resolves all references in `components` before resolving
them in the rest of the spec to ensure they are always done in the
proper order.